+++ src/components/forum/PostPage.jsx
@@ -1,21 +1,32 @@
 import axios from 'axios'
 import { useState, useEffect } from 'react'
-import { useSelector } from 'react-redux'
+import { useSelector, useDispatch } from 'react-redux'
 import Comment from './Comment.jsx'
-import { useParams } from 'react-router-dom'
+import { useParams, useNavigate } from 'react-router-dom'
 import { AiOutlineLike, AiFillLike } from "react-icons/ai"
 
 export default function PostPage() {
     const { postID } = useParams()
+
+    //state variables
     const [post, setPost] = useState({})
     const [comments, setComments] = useState([])
     const [author, setAuthor] = useState('')
     const [likes, setLikes] = useState(0)
-    const [isLiked, setIsLiked] = useState(true)
+    const [isLiked, setIsLiked] = useState(false)
     const [postLikeID, setPostLikeID] = useState(null)
     const [body, setBody] = useState('')
     const [isEditing, setIsEditing] = useState(false)
+    const [page, setPage] = useState(1)
+    const [totalPages, setTotalPages] = useState(1)
+
+    //redux store variables
     const userID = useSelector((state) => state.userID)
+    const loggedIn = useSelector((state) => state.loggedIn)
+
+    //action declarations
+    const dispatch = useDispatch()
+    const navigate = useNavigate()
 
     const getPost = async () => {
         const res = await axios.get(`/api/posts/${postID}`) 
@@ -27,27 +38,54 @@ export default function PostPage() {
     }
 
     const getComments = async () => {
-        const res = await axios.get(`/api/postComments/${postID}`)
+        const res = await axios.get(`/api/postComments/${postID}`, {
+            params: { page: page }
+        })
         setComments(res.data.comments)
+        setTotalPages(res.data.totalPages)
+    }
+
+    const sessionCheck = async () => {
+        const res = await axios.get('/api/session-check')
+
+        if (res.data.success) {
+            dispatch({
+                type: "USER_AUTH",
+                payload: {
+                    userID: res.data.userID,
+                    username: res.data.username,
+                    loggedIn: true
+                }
+            })
+        }
+        else {
+            navigate("/login")
+        }
     }
 
     const toggleLike = async () => {
-        if (isLiked) {
-            await axios.delete(`/api/postLikes/${postLikeID}`)
-            setLikes(likes - 1)
-            setIsLiked(false)
-        }
-        else {
-            const body = { userID: userID}
-            const res = await axios.post(`/api/postLikes/${postID}`, body)
-            setLikes(likes + 1)
-            setIsLiked(true)
-            setPostLikeID(res.data.postLikeID)
+        sessionCheck()
+        if (loggedIn) {
+            if (isLiked) {
+                await axios.delete(`/api/postLikes/${postLikeID}`)
+                setLikes(likes - 1)
+                setIsLiked(false)
+            }
+            else {
+                const body = { userID: userID}
+                const res = await axios.post(`/api/postLikes/${postID}`, body)
+                setLikes(likes + 1)
+                setIsLiked(true)
+                setPostLikeID(res.data.postLikeID)
+            }
         }
     }
 
     const createComment = () => {
-        setIsEditing(true)
+        sessionCheck()
+        if (loggedIn) {
+            setIsEditing(true)
+        }
     }
 
     const handleSubmit = async (event) => {
@@ -71,7 +109,7 @@ export default function PostPage() {
     useEffect(() => {
         getPost();
         getComments();
-    }, [])
+    }, [page])
 
     return (
         <div>
@@ -97,6 +135,15 @@ export default function PostPage() {
                 :
                 <span></span>
             }
+
+            {/* Pagination controls */}
+            {totalPages > 1 && (
+                <div>
+                    {Array.from({ length: totalPages }, (_, index) => (
+                        <button key={index} onClick={() => setPage(index + 1)}>{index + 1}</button>
+                    ))}
+                </div>
+            )}
         </div>
     )
 }
